requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
psycopg2-binary>=2.9.0
pg8000>=1.30.0
python-dotenv>=1.0.0